dev_performance_dashboard/requirements.txt
streamlit
pandas
plotly
PyGithub
invoke
pycodestyle
pydocstyle
PyGithub
pytest